pyproject.toml
[tool.poetry]
name = "newsflow-bot"
version = "0.1.0"
description = "Self-hosted RSS to Discord/Telegram bot with translation support"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "newsflow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# === Core Dependencies ===
# Messaging platforms
"discord.py" = "^2.3.0"
python-telegram-bot = { extras = ["job-queue"], version = "^20.7" }

# RSS parsing
feedparser = "^6.0.0"
aiohttp = "^3.9.0"

# Database
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
aiosqlite = "^0.19.0"
alembic = "^1.13.0"

# Configuration
pydantic-settings = "^2.1.0"

# HTML processing
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"

# Scheduling
apscheduler = "^3.10.0"

# Date parsing
python-dateutil = "^2.8.0"

# Logging
structlog = "^24.1.0"

# === Optional Dependencies ===
# API service
fastapi = { version = "^0.109.0", optional = true }
uvicorn = { extras = ["standard"], version = "^0.27.0", optional = true }

# Translation services
google-cloud-translate = { version = "^3.14.0", optional = true }
deepl = { version = "^1.16.0", optional = true }
openai = { version = "^1.10.0", optional = true }

# Cache
redis = { extras = ["hiredis"], version = "^5.0.0", optional = true }

# PostgreSQL
asyncpg = { version = "^0.29.0", optional = true }

[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
translation-google = ["google-cloud-translate"]
translation-deepl = ["deepl"]
translation-openai = ["openai"]
translation = ["google-cloud-translate", "deepl"]
cache = ["redis"]
postgres = ["asyncpg"]
all = [
    "fastapi", "uvicorn",
    "google-cloud-translate", "deepl", "openai",
    "redis", "asyncpg"
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
newsflow = "newsflow.main:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
